pharma-automation/cloud/requirements.txt
fastapi>=0.115
uvicorn[standard]
sqlalchemy[asyncio]>=2.0
asyncpg
alembic
pydantic>=2.0
pydantic-settings
python-multipart
httpx
bcrypt>=4.0
PyJWT>=2.8
slowapi>=0.1.9
pytest
pytest-asyncio
psycopg2-binary
